pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "chromatintrek"
version = "0.1.0"
description = "End-to-end pipeline for ATAC-seq, ChIP-seq, and CUT&Tag chromatin analysis"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "ChromatinTrek" }]
requires-python = ">=3.9"
keywords = ["bioinformatics", "atac-seq", "chip-seq", "cut-and-tag", "chromatin", "epigenomics"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "seaborn>=0.12",
  "pyyaml>=6.0",
  "requests>=2.31",
  "tqdm>=4.65",
  "joblib>=1.3",
  "pybedtools>=0.9",
  "pyranges>=0.0.129",
  "gseapy>=1.0",
  "logomaker>=0.8",
]

[project.optional-dependencies]
dev = ["pytest", "black", "flake8", "isort"]

[project.scripts]
chromatintrek = "chromatintrek.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["chromatintrek*"]

[tool.black]
line-length = 99
target-version = ["py39", "py310", "py311"]
